Ifpack2: Fix a snafu I made with a size_t in a parameter list. Add UserPartitioner to the Ifpack2::Details namespace.
There's no reason the parameter "partitioner: overlap" needs to be a size_t.
Thus, I changed it to an int, which meant changing it in the Partitioner header
and the OverlappingPartitioner. I also deleted a bunch of junk from my little
test/vanka directory. The UserPartitioner that I copied over from Ifpack was
living there while I was playing with it; now that that is done, it got moved
to the ifpack2/src directory and put in the Ifpack2::Details namespace.

Update, re: Blocked Gauss-Seidel/Additive Schwarz. In my own head, I've
downgraded it from a bug to a strong disagreement. It "works" in the sense
of doing something not-too-unreasonable, but still not "correct" in the
sense that I want it to be. Still a work in progress. But it passes the
testing, so I guess that's good enough for a commit.

Build/Test Cases Summary
Enabled Packages: Ifpack2
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=22,notpassed=0 (18.46 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (6.57 min)
Other local commits for this build/test group: 4f3e5eb, f57609e, de7e0fd

+++ packages/ifpack2/src/Ifpack2_OverlappingPartitioner_decl.hpp
@@ -92,7 +92,7 @@ public:
   LocalOrdinal numLocalParts() const;
 
   //! Returns the overlapping level.
-  size_t overlappingLevel() const;
+  int overlappingLevel() const;
 
   //! Returns the local non-overlapping partition ID of the specified row.
   /*! Returns the non-overlapping partition ID of the specified row.
@@ -170,7 +170,7 @@ protected:
   //! Reference to the graph to be partitioned
   Teuchos::RCP<const Tpetra::RowGraph<LocalOrdinal,GlobalOrdinal,Node> > Graph_;
   //! Overlapping level.
-  size_t OverlappingLevel_;
+  int OverlappingLevel_;
   //! If \c true,  the graph has been successfully partitioned.
   bool IsComputed_;
   //! If \c true, information are reported on cout.
